refactor(services): move GitHub-related functions to separate module
Consolidate all GitHub API-related functions (`getMarkers`, `getItems`, `getMapNames`) into a dedicated `github.ts` module to improve code organization and maintainability. This separation clarifies the responsibility of each module and reduces clutter in the `services.ts` file.
