+++ src/components/ClanMaps/CreateMapNoLog.tsx
@@ -1,7 +1,7 @@
 import type React from "react";
 import { useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { getMapNames } from "@functions/services";
+import { getMapNames } from "@functions/github";
 import CreateMapPanel from "./CreateMapPanel";
 import { getDomain } from "@functions/utils";
 import { addMap } from "@functions/requests/maps";
