+++ src/pages/Crafter.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect, useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { Helmet } from "react-helmet";
 import queryString from "query-string";
-import { getItems } from "@functions/services";
+import { getItems } from "@functions/github";
 import ModalMessage from "@components/ModalMessage";
 import Items from "@components/Crafter/Items";
 import SelectedItem from "@components/Crafter/SelectedItem";
