+++ src/pages/TechTree.tsx
@@ -8,7 +8,7 @@ import React, {
 } from "react";
 import { useTranslation } from "react-i18next";
 import { NavLink, useParams } from "react-router";
-import { getItems, getStoredItem, storeItem } from "@functions/services";
+import { getStoredItem, storeItem } from "@functions/services";
 import { useUser } from "@store/userStore";
 import LoadingScreen from "@components/LoadingScreen";
 import ModalMessage from "@components/ModalMessage";
@@ -18,6 +18,7 @@ import { getLearned, addTech, getUser } from "@functions/requests/users";
 import HeaderMeta from "@components/HeaderMeta";
 import type { Item } from "@ctypes/item";
 import { Tree } from "@ctypes/dto/tech";
+import { getItems } from "@functions/github";
 
 const SkillTreeTab = React.lazy(
   () => import("@components/TechTree/SkillTreeTab"),
